backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.2
flask-cors==6.0.1